% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{avg_pe}
\alias{avg_pe}
\title{Average P/E}
\usage{
avg_pe(w, x)
}
\arguments{
\item{w}{weight vector}

\item{x}{P/E vector}
}
\value{
weighted average P/E
}
\description{
Average P/E
}
\details{
Filters out missing and negative values and re-weights
}
