% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_time_series.R
\name{calc_drawdown}
\alias{calc_drawdown}
\title{Calculate Peak to Trough Drawdowns}
\usage{
calc_drawdown(x)
}
\arguments{
\item{x}{xts object}
}
\value{
xts of drawdown time-series
}
\description{
Calculate Peak to Trough Drawdowns
}
