% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_time_series.R
\name{calc_uni_beta}
\alias{calc_uni_beta}
\title{Calculate Univariate Beta}
\usage{
calc_uni_beta(mgr, bench, rf)
}
\arguments{
\item{mgr}{xts manager(s)}

\item{bench}{xts benchmark}

\item{rf}{xts risk-free}
}
\value{
vector of betas to benchmark
}
\description{
Calculate Univariate Beta
}
