% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{clean_emv}
\alias{clean_emv}
\title{Standardize EMV column header for $ value}
\usage{
clean_emv(df)
}
\description{
Standardize EMV column header for $ value
}
\details{
holdings files from SEC will have `valUSD` as column header or field
  representing value, change to `emv` to standardize with Black Diamond pull
}
