% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_bd}
\alias{download_bd}
\title{Download Holdings from BlackDiamond}
\usage{
download_bd(account_id, api_keys, as_of = NULL)
}
\arguments{
\item{account_id}{string to represent target account (internal ID)}

\item{as_of}{date to pull holdings as of, latest is last business day}
}
\value{
data.frame with holdings
}
\description{
Download Holdings from BlackDiamond
}
