% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_fs}
\alias{download_fs}
\title{Factset Formula API}
\usage{
download_fs(api_keys, ids, formulas, type = c("ts", "cs"), flatn = TRUE)
}
\arguments{
\item{api_keys}{list with API keys}

\item{ids}{string vector with ids, e.g., ticker, CUSIP, ISIN}

\item{formulas}{string vector with formulas, see details}

\item{features}{string vector describing formula results, see details}
}
\value{
data.frame with data
}
\description{
Factset Formula API
}
\details{
Formulas are factset API formulas, e.g., "P_PRICE(0, -2, D)". You
  can list multiple formulas as strings:
  c("P_PRICE(0, -2, D)", "FF_EPS(QTR_R, 0)"). Features are a description to
  match results of formulas, in this example c("Price", "EPS").
}
