% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_fs_ctf_ret}
\alias{download_fs_ctf_ret}
\title{Download Daily Returns from Factset}
\usage{
download_fs_ctf_ret(id, api_keys, t_minus = 5, freq = "D")
}
\arguments{
\item{id}{factset ID, e.g., "CLIENT:/PA_SOURCED_RETURNS/DTC_US_ACTIVE_EQUITY_COMMON_FUND"}

\item{api_keys}{list with API keys}

\item{t_minus}{integer for how many days back to start download}
}
\value{
list with id and xts of returns
}
\description{
Download Daily Returns from Factset
}
