% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_fs_gp}
\alias{download_fs_gp}
\title{Download from Factset Global Prices API}
\usage{
download_fs_gp(api_keys, ids, date_start, date_end, freq = "D")
}
\description{
Download from Factset Global Prices API
}
