% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_sec_nport}
\alias{download_sec_nport}
\title{Download Holdings from SEC Database}
\usage{
download_sec_nport(long_cik, short_cik, user_email)
}
\arguments{
\item{long_cik}{string to represent the fund / share-class CIK,
typically longer in length}

\item{short_cik}{string to represent the parent company CIK,
typically shorter in length}

\item{user_email}{SEC's API requires an email address}

\item{as_of}{optional date to mark when the download occurred}
}
\value{
data.frame with fund holdings
}
\description{
Download Holdings from SEC Database
}
