% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_port.R
\name{dtc_name_match_ret}
\alias{dtc_name_match_ret}
\title{Match DTCNames to daily and monthly returns}
\usage{
dtc_name_match_ret(dtc_name, ret, force_month = FALSE)
}
\arguments{
\item{dtc_name}{DTCNames to match}

\item{ret}{list of daily and monthly returns from Database}

\item{force_month}{only look for monthly returns}
}
\description{
Match DTCNames to daily and monthly returns
}
\details{
Daily returns are matched first. If daily returns are missing then the
monthly return files are checked. If there is a combo of daily and monthly
the daily returns are converted to monthly returns and combined.
}
