% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_time_series.R
\name{excess_ret}
\alias{excess_ret}
\title{Calculate excess return (a - b)}
\usage{
excess_ret(a, b)
}
\arguments{
\item{a}{xts object to transform into excess return}

\item{b}{univariate xts object of benchmark or risk-free rate to subtract
from `a`}
}
\description{
Calculate excess return (a - b)
}
