% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{group_df}
\alias{group_df}
\title{Group holdings data.frame}
\usage{
group_df(ddf, gp_by, bucket, bdf = NULL, recons_wgt = TRUE)
}
\arguments{
\item{ddf}{holdings data.frame that has been drilled down, see
drill_down}

\item{gp_by}{string to represent which column to group data by}

\item{bdf}{benchmark data.frame}

\item{recons_wgt}{boolean for weights to sum to 1}
}
\description{
Group holdings data.frame
}
