% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{hfr_asci_to_s3}
\alias{hfr_asci_to_s3}
\title{Read HFR ASCI file and save to S3}
\usage{
hfr_asci_to_s3(fpath, bucket, s3_file_name)
}
\arguments{
\item{fpath}{file path location of saved ASCI files}

\item{bucket}{S3 File System from Database}

\item{s3_file_name}{prefix of name to save, e.g., "hfri-all"}
}
\description{
Read HFR ASCI file and save to S3
}
\details{
Download the ASCI file from the HFRI database download section
The fpath will contain multiple files, the returns and meta data file
will be read in and saved. The s3_file_name parameter will determine the names,
e.g., "hfri-all" will save "hfri-all-ror.parquet" and "hfri-all-meta.parquet"
to S3.
}
