% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{hfr_csv_to_s3}
\alias{hfr_csv_to_s3}
\title{Save HFR ROR data to S3}
\usage{
hfr_csv_to_s3(bucket, s3_file_name, wb = NULL, dat = NULL)
}
\arguments{
\item{bucket}{s3 bucket from arrow}

\item{s3_file_name}{name to save in s3}

\item{wb}{optional file path to load csv file}

\item{dat}{optional data.frame or tibble to upload (instead of reading csv)}
}
\value{
nothing is returned, if successful the file will be uploaded to S3
}
\description{
Save HFR ROR data to S3
}
