% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{id_match_msl}
\alias{id_match_msl}
\title{Match ids to MSL}
\usage{
id_match_msl(ids, msl)
}
\arguments{
\item{ids}{ids to match}

\item{msl}{Master Security List from Database}
}
\description{
Match ids to MSL
}
\details{
ids are searched in the following order: DTCName, Ticker, ISIN, CUSIP, SEDOL,
LEI, and Identifier
}
