% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_error_hndl.R
\name{is_null_dim}
\alias{is_null_dim}
\title{Utility function to check if an object has dimensions}
\usage{
is_null_dim(x)
}
\value{
TRUE if it has dimensions, FALSE if not
}
\description{
Utility function to check if an object has dimensions
}
