% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{left_merge_msl}
\alias{left_merge_msl}
\title{Left join master security list}
\usage{
left_merge_msl(mdf, msl)
}
\arguments{
\item{mdf}{list returned from `merge_msl`, see details}

\item{msl}{master security list}
}
\value{
list with all, matched, and missing data
}
\description{
Left join master security list
}
\details{
This function is designed to rejoin data to the MSL where some of
  the data has been joined and some has not, for example in when drilling
  down in holdings the holdings will not have been joined but higher level
  securities will have the joined data
}
