% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{match_ret_df}
\alias{match_ret_df}
\title{Match returns to holdings}
\usage{
match_ret_df(mdf, ret)
}
\arguments{
\item{ret}{list of returns, ret$d = daily, ret$m = monthly}
}
\description{
Match returns to holdings
}
