% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{merge_country}
\alias{merge_country}
\title{Merge country data}
\usage{
merge_country(ddf, bucket)
}
\arguments{
\item{ddf}{list output from drill_down or merge_msl for a group of stocks}

\item{bucket}{S3 file system from Database Object or arrow}
}
\value{
ddf list with merged sectors in ddf$match data.frame
}
\description{
Merge country data
}
