% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{merge_list}
\alias{merge_list}
\title{Merge list with data.frames containing 2 columns}
\usage{
merge_list(x)
}
\arguments{
\item{x}{list of data.frames}
}
\value{
data.frame with merged data
}
\description{
Merge list with data.frames containing 2 columns
}
\details{
assumes key ids are in column 1 and the data you want to merge
is in column 2
}
