% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{merge_macro}
\alias{merge_macro}
\title{Merge Piper Sandler Macro Select Model Rankings}
\usage{
merge_macro(ddf, macro_df)
}
\arguments{
\item{ddf}{list output from drill_down or merge_msl for a group of stocks}

\item{macro_df}{data.frame of macro rankings, e.g., db$macro$r3, see
update_macro}
}
\description{
Merge Piper Sandler Macro Select Model Rankings
}
\note{
Relies on column order given names of factors will change, will need to
  update if columns are added or deleted from the piper sandler workbook
}
