% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{merge_msl}
\alias{merge_msl}
\title{Merge MSL and holdings data.frame (df)}
\usage{
merge_msl(df, msl)
}
\arguments{
\item{df}{holdings data.frame}

\item{msl}{master security list}
}
\value{
list with all, match, and miss to represent union, intersection, and
  union - intersection
}
\description{
Merge MSL and holdings data.frame (df)
}
