% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{msl_match}
\alias{msl_match}
\title{Match securities in a portfolio with the Master Security List}
\usage{
msl_match(df, msl)
}
\arguments{
\item{df}{data.frame of securities, see details}

\item{msl}{Master Security List}
}
\value{
numeric vector with row numbers of the match in the `msl`
}
\description{
Match securities in a portfolio with the Master Security List
}
\details{
The match will attempt to match the following `df` columns
  ticker, cusip, name, isin, lei, and identifier
}
