% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{port_from_ids}
\alias{port_from_ids}
\title{Create a Portfolio from ids}
\usage{
port_from_ids(msl, ret, ids, wgt = NULL, nm = "Port", reb_freq = "M")
}
\arguments{
\item{msl}{master security list from Database}

\item{ret}{return list from Database}

\item{wgt}{optional weight vector for portfolio}

\item{nm}{optional string for portfolio name}

\item{reb_freq}{optional character to specify rebalance frequency}
}
\description{
Create a Portfolio from ids
}
