% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{Portfolio}
\alias{Portfolio}
\title{Portfolio Object}
\description{
Investment portfolio
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{name of portfolio for display}

\item{\code{asset_ret}}{xts of asset returns}

\item{\code{asset_wgt}}{xts of weights for corresponding asset returns}

\item{\code{port_ret}}{xts of portfolio returns as a result of rebalancing}

\item{\code{port_wealth}}{xts of cumulative portfolio returns}

\item{\code{perf_stats}}{list with performance reports}

\item{\code{reb_stats}}{list with rebalance stats for attribution}

\item{\code{risk_stats}}{list multi-variate risk stats}

\item{\code{opt}}{list with optimization output}

\item{\code{ret_freq}}{return frequency}

\item{\code{reb_freq}}{optional field for automatic rebalancing frequency}

\item{\code{reb_wgt}}{xts of rebalance weights}

\item{\code{rf}}{xts of risk-free time-series}

\item{\code{benchmark}}{optional field for benchmark (also a Portfolio Object)}

\item{\code{sum_to_1}}{boolean to re-calculate weights to pro-rata sum to 1 if
any returns or weights are missing}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sum_to_1}}{boolean to re-calculate weights to pro-rata sum to 1 if
any returns or weights are missing}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Portfolio-new}{\code{Portfolio$new()}}
\item \href{#method-Portfolio-clean_asset_ret}{\code{Portfolio$clean_asset_ret()}}
\item \href{#method-Portfolio-align_reb_wgt}{\code{Portfolio$align_reb_wgt()}}
\item \href{#method-Portfolio-check_wgt_ret}{\code{Portfolio$check_wgt_ret()}}
\item \href{#method-Portfolio-lazy_reb_wgt}{\code{Portfolio$lazy_reb_wgt()}}
\item \href{#method-Portfolio-rebal}{\code{Portfolio$rebal()}}
\item \href{#method-Portfolio-clone}{\code{Portfolio$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-new"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$new(
  name = "port",
  asset_ret = NULL,
  ret_freq = c("D", "M", "Q", "A"),
  reb_freq = c(NULL, "D", "M", "Q", "A"),
  reb_wgt = NULL,
  rf = NULL,
  benchmark = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-clean_asset_ret"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-clean_asset_ret}{}}}
\subsection{Method \code{clean_asset_ret()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$clean_asset_ret(asset_ret = NULL, eps = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-align_reb_wgt"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-align_reb_wgt}{}}}
\subsection{Method \code{align_reb_wgt()}}{
Align rebalance weights with corresponding asset returns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$align_reb_wgt(sum_to_1 = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Weights are either a named vector or xts time-series. The alignment
  will look for the intersection of names. Any names not found in the
  intersection will be removed. If sum_to_1 is set to TRUE the weights will
  be recalculated to sum to 100%. If the vector is not named a
  warning will be issued and the length of the weights and number of
  columns of the returns will be checked.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-check_wgt_ret"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-check_wgt_ret}{}}}
\subsection{Method \code{check_wgt_ret()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$check_wgt_ret()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-lazy_reb_wgt"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-lazy_reb_wgt}{}}}
\subsection{Method \code{lazy_reb_wgt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$lazy_reb_wgt(reb_wgt = NULL, reb_freq = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-rebal"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-rebal}{}}}
\subsection{Method \code{rebal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$rebal(
  asset_ret = NULL,
  reb_wgt = NULL,
  align_wgt = TRUE,
  sum_to_1 = TRUE,
  clean_asset_ret = TRUE,
  eps = 0.05
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Portfolio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Portfolio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Portfolio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
