% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_hfr_asci}
\alias{read_hfr_asci}
\title{Read HFR ASCI file and return a list with data.tables of returns
and meta-data}
\usage{
read_hfr_asci(fpath)
}
\arguments{
\item{fpath}{folder location of downloaded files}
}
\description{
Read HFR ASCI file and return a list with data.tables of returns
and meta-data
}
