% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{read_hfr_csv}
\alias{read_hfr_csv}
\title{Read HFR ROR csv file}
\usage{
read_hfr_csv(wb, min_track = 30)
}
\arguments{
\item{wb}{file path of csv}

\item{min_track}{length (months) of minimum track record, set to 0 to not
exclude any funds}
}
\value{
tibble of returns in tidy longer format to preserve meta-data
}
\description{
Read HFR ROR csv file
}
