% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{read_macro_wb}
\alias{read_macro_wb}
\title{Read Piper Sandler Macro Select Workbook}
\usage{
read_macro_wb(wb, idx, fact_nm)
}
\arguments{
\item{wb}{workbook full file name including path}

\item{idx}{string representing which index to use, e.g., "Russell 3000"}

\item{fact_nm}{string vector representing names of the current macro factors}
}
\description{
Read Piper Sandler Macro Select Workbook
}
