% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_holdings.R
\name{split_by_n}
\alias{split_by_n}
\title{Split data.frame into n groups from a numeric field}
\usage{
split_by_n(x, tgt, n)
}
\arguments{
\item{x}{data.frame}

\item{tgt}{target column name to form split on}

\item{n}{integer number of groups to form}
}
\value{
list of split data.frames
}
\description{
Split data.frame into n groups from a numeric field
}
\details{
Groups are sorted in ascending order by quantile function. Missing
values are removed.
}
